[lldb] Move LLDBMemoryReader to its own file
(cherry picked from commit 55ecb91fc45c3fd8bf59f5514db121643b5e1df8)
