+++ lldb/source/Plugins/LanguageRuntime/Swift/CMakeLists.txt
@@ -1,4 +1,5 @@
 add_lldb_library(lldbPluginSwiftLanguageRuntime PLUGIN
+  LLDBMemoryReader.cpp
   SwiftLanguageRuntime.cpp
   SwiftLanguageRuntimeDynamicTypeResolution.cpp
   SwiftLanguageRuntimeNames.cpp
